requirements.txt
pytz==2018.7
Scrapy==1.5.1
python-dateutil==2.7.5